requirements.txt
aiofiles==23.2.1
annotated-types==0.6.0
anyio==3.7.1
asyncio==3.4.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.105.0
filelock==3.13.1
fsspec==2023.12.2
h11==0.14.0
httptools==0.6.1
idna==3.6
Jinja2==3.1.2
llvmlite==0.41.1
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
openai-whisper @ git+https://github.com/openai/whisper.git@8bc8860694949db53c42ba47ddc23786c2e02a8b
Pillow==10.1.0
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tiktoken==0.5.2
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
